Move RewardsTransactionTableController close to RewardsTransactionTable.
